+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Expand `set_partial_values` tests
  - Specialise `set_partial_values` for `MemoryStore`
  - Bump maximum supported `ndarray` version from 0.15 to 0.16
+ - **Breaking**: Make `create_chunk_grid_{regular,rectangular}` `pub(crate)` in alignment with other internal create from metadata methods
 
 ### Fixed
  - `[async_]store_set_partial_values` no longer truncates